refactored copy buffer
